Group text encoding related code
